client/backend/pyproject.toml
[project]
name = "racksmith-client"
version = "0.1.0"
description = "Racksmith client — Ansible automation API and background worker"
readme = "README.md"
requires-python = ">=3.14"
license = {file = "../../LICENSE"}
authors = [{name = "Ekku Leivonen"}]

dependencies = [
    "ansible-core>=2.20.3",
    "arq>=0.25.0",
    "asyncssh>=2.22.0",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "mac-vendor-lookup>=0.1.15",
    "openai>=2.26.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "redis>=7.2.1",
    "ruamel-yaml>=0.19.1",
    "slowapi>=0.1.9",
    "structlog>=25.5.0",
]

[project.urls]
Homepage = "https://github.com/ekkuleivonen/racksmith"
Repository = "https://github.com/ekkuleivonen/racksmith"
Issues = "https://github.com/ekkuleivonen/racksmith/issues"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pip-audit>=2.10.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.0",
    "ruff>=0.15.5",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-p no:debugging -p no:faulthandler"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = false
warn_unused_ignores = true
strict_optional = true

